+++ gnark/gnark-jni/gnark-jni.go
@@ -15,13 +15,14 @@ func MiMCHash(b []byte) []byte {
 }
 
 
-//export compute
-func compute(input *C.char, inputLength C.int, output *C.char) C.int {
+//export computeMimc
+func computeMimc(input *C.char, inputLength C.int, output *C.char) C.int {
        inputSlice := C.GoBytes(unsafe.Pointer(input), inputLength)
        outputSlice := (*[32]byte)(unsafe.Pointer(output))[:]
        hash := MiMCHash(inputSlice)
        copy(outputSlice, hash)
        return C.int(len(hash))
 }
+
 
 func main() {}
