+++ gnark/src/test/java/org/hyperledger/besu/nativelib/gnark/LibGnarkTest.java
@@ -13,7 +13,7 @@
  * SPDX-License-Identifier: Apache-2.0
  *
  */
-package org.hyperledger.besu.nativelib.mimc;
+package org.hyperledger.besu.nativelib.gnark;
 
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
@@ -22,19 +22,19 @@ import org.junit.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-public class LibMimcTest {
+public class LibGnarkTest {
 
     @Test
     public void testHashZero() {
         byte[] output = new byte[Bytes32.SIZE];
-        LibMimc.compute(Bytes32.ZERO.toArrayUnsafe(), Bytes32.SIZE, output);
+        LibGnark.computeMimc(Bytes32.ZERO.toArrayUnsafe(), Bytes32.SIZE, output);
         assertThat(Bytes.wrap(output)).isEqualTo(Bytes.fromHexString("0x2c7298fd87d3039ffea208538f6b297b60b373a63792b4cd0654fdc88fd0d6ee"));
     }
 
     @Test
     public void testHashOne() {
         byte[] output = new byte[Bytes32.SIZE];
-        LibMimc.compute(Bytes32.leftPad(Bytes.of(1)).toArrayUnsafe(), Bytes32.SIZE, output);
+        LibGnark.computeMimc(Bytes32.leftPad(Bytes.of(1)).toArrayUnsafe(), Bytes32.SIZE, output);
         assertThat(Bytes.wrap(output)).isEqualTo(Bytes.fromHexString("0x27e5458b666ef581475a9acddbc3524ca252185cae3936506e65cda9c358222b"));
     }
 
@@ -45,7 +45,7 @@ public class LibMimcTest {
             input.set(Bytes32.SIZE*(i+1)-1,(byte) i);
         }
         byte[] output = new byte[Bytes32.SIZE];
-        LibMimc.compute(input.toArrayUnsafe(), input.size(), output);
+        LibGnark.computeMimc(input.toArrayUnsafe(), input.size(), output);
         assertThat(Bytes.wrap(output)).isEqualTo(Bytes.fromHexString("0x145875dd085ea2fb9796333e55c9da80228eb321df0ca9a41ca64ba6fe90b167"));
     }
 }
