+++ settings.gradle
@@ -21,4 +21,4 @@ include 'bls12-381'
 include 'ipa-multipoint'
 include 'secp256k1'
 include 'secp256r1'
-include 'mimc'
+include 'gnark'
